requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.2
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
pydantic==2.4.2
pydantic-core==2.10.1
python-dotenv==1.0.0
python-telegram-bot==20.6
PyYAML==6.0.1
requests==2.31.0
sarufi==0.1.9
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.8.0
urllib3==2.1.0
python-multipart==0.0.6
mangum==0.17.0

